btlx and jackraftercut unittest separation
